Create README.md
How to use these files
The attached files do NOT work as a standalone project.
Instead, you should use them to compare my code (in the lib/ folder, as well as the 
pubspec.yaml file) to your code. Experience shows, that this helps you spot + fix 
99% of all errors you might encounter.
If you want to run the code, you can do that as well though. 
Follow these steps
1) Create a new Flutter project via flutter create some_project
2) Replace the lib/ folder and the pubspec.yaml file with the attached files
3) Possibly add assets like images or fonts or adjust the AndroidManifest.xml / 
Info.plist file to add data/ permissions (only if covered in the previous lectures)
4) flutter packages get
5) flutter run (with a running emulator or real device connected to your machine)
to view the app